Added experimental features for multiple objects
Added the overload of ',' operator to create a list of ThreadSafe objects and lock them for the duration of the full expression where the list is present. Added a new type (Locks) to hold the list of locked mutexes. Added the declaration of '->*' operator to call a function after the lockdown of one or more ThreadSafe objects. Debug of each of these new features is strongly needed