Made some experiments
I tried to make som experiments to overload in the right way << operator. I couldn't find a way. The issue is that after the call to * operator I always have to deal with an rvalue, so I cannot distinghish between a atd::move(d) argument and the argument returned by * operator (because overload resolution doesn't prefer rvalue references over rvalues).